feat: add start page with navigation and rename start to signin
