+++ src/components/Button/Button.scss
@@ -26,4 +26,13 @@
         background-color: $color-light;
         color: $color-main-first;
     }
+    &_navigate {
+        background-color: $color-light;
+        border: 1px solid $color-main-first;
+        color: $color-main-first;
+        font-weight: 500;
+        &:hover {
+            box-shadow: 0px 0px 6px rgba(0, 41, 128, 20%);
+        }
+    }
 }
