+++ src/pages/signin/signin.scss
@@ -1,0 +1,35 @@
+.login-form {
+  max-width: 400px;
+  width: 30%;
+  min-height: 60%;
+  height: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-around;
+  box-shadow: 0px 0px 6px rgba(0, 0, 0, 14%);
+  border-radius: 12px;
+  row-gap: 40px;
+  padding: 20px;
+  margin: auto;
+
+  &__group {
+    width: 90%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-around;
+    row-gap: 20px;
+    padding: 10px 0;
+  }
+
+  &__bottom {
+    width: 60%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-around;
+    row-gap: 20px;
+    padding: 10px 0;
+  }
+}
